1、将 panic 直接输出错误。 2、加入部分发短信相关逻辑代码。

+++ http/sms/controller.go
@@ -24,7 +24,7 @@ func PostCode(ctx *gin.Context) {
 
 	count := countByPhoneAndDay(req.Phone, makeToday())
 	if count > 5 {
-		dataMap := make(map[string]int)
+		dataMap := make(map[string]int64)
 		dataMap["count"] = count
 
 		app.Response(ctx, result.RateLimit.WithData(dataMap))
